pyproject.toml
[tool.poetry]
name = "dentora"
version = "0.1.0"
description = "Backend for dental appointment management"
authors = ["Fede <federicogacastro@gmail.com>"]
readme = "README.md"
packages = [{include = "dentora"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15"
psycopg2-binary = "^2.9"
celery = {extras = ["redis"], version = "^5.3"}
django-environ = "^0.11"
drf-spectacular = "^0.27"
djangorestframework-simplejwt = "^5.3"
gunicorn = "^22.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-django = "^4.8"
pytest-cov = "^5.0"
factory-boy = "^3.3"
black = "^24.0"
flake8 = "^7.0"
mypy = "^1.9"
django-stubs = {extras = ["compatible-mypy"], version = "^5.0"}
djangorestframework-stubs = {extras = ["compatible-mypy"], version = "^3.15"}

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "dentora.*.tests.*"
ignore_errors = true

[[tool.mypy.overrides]]
# django-environ ships no type stubs
module = "environ"
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "dentora.config.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "dentora.config.settings"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--strict-markers"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
